requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Werkzeug==2.3.7
requests==2.29.0
python-dotenv==1.0.0
google-generativeai==0.3.0
reportlab==4.0.4
Pillow==10.0.1
firebase-admin==6.2.0
pyrebase4==4.7.1
gunicorn==21.2.0
waitress==2.1.2
